app/requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.0
Flask==2.3.2
future==0.18.3
h11==0.14.0
httpcore==0.17.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pyadaptivecards==0.1.1
PyJWT==2.8.0
pymongo==4.4.1
pytest==7.3.1
python-dotenv==1.0.0
requests==2.30.0
requests-toolbelt==1.0.0
sniffio==1.3.0
urllib3==2.0.4
webexteamssdk==1.6.1
Werkzeug==2.3.6